Refactor gradle plugin packages
