+++ scabbard-gradle-plugin/src/main/kotlin/dev/arunkumar/scabbard/gradle/processor/processor.kt
@@ -4,12 +4,26 @@ import dev.arunkumar.scabbard.gradle.ScabbardGradlePlugin.Companion.ANDROID_APP_
 import dev.arunkumar.scabbard.gradle.ScabbardGradlePlugin.Companion.ANDROID_LIBRARY_PLUGIN_ID
 import dev.arunkumar.scabbard.gradle.ScabbardGradlePlugin.Companion.JAVA_LIBRARY_PLUGIN_ID
 import dev.arunkumar.scabbard.gradle.ScabbardGradlePlugin.Companion.KAPT_PLUGIN_ID
+import dev.arunkumar.scabbard.gradle.ScabbardGradlePlugin.Companion.SCABBARD_PLUGIN_ID
 import org.gradle.api.Project
 import org.gradle.api.artifacts.Configuration
+import org.gradle.api.artifacts.ConfigurationContainer
 import org.gradle.api.artifacts.Dependency
 
 internal const val KAPT_CONFIG = "kapt"
 internal const val ANNOTATION_PROCESSOR_CONFIG = "annotationProcessor"
+internal const val SCABBARD_GROUP = "dev.arunkumar"
+internal const val SCABBARD_NAME = "scabbard-processor"
+internal const val SCABBARD_PROCESSOR_FORMAT = "$SCABBARD_GROUP:$SCABBARD_NAME:%s"
+
+private fun Dependency.isScabbardDependency() =
+  group == SCABBARD_GROUP && name == SCABBARD_NAME
+
+@Suppress("unused")
+private fun Dependency.isDaggerCompiler() =
+  group == "com.google.dagger" && name == "dagger-compiler"
+
+private fun Configuration.hasScabbard() = dependencies.any(Dependency::isScabbardDependency)
 
 private fun configName(isKapt: Boolean) = when {
   isKapt -> KAPT_CONFIG
@@ -31,7 +45,32 @@ private fun Project.removeScabbard(isKapt: Boolean = false) {
     }
 }
 
-private fun Configuration.hasScabbard() = dependencies.any(Dependency::isScabbardDependency)
+internal class VersionCalculator(private val project: Project) {
+
+  companion object {
+    private const val CLASSPATH_CONFIG = "classpath"
+  }
+
+  private fun versionFromConfigurations(configurations: ConfigurationContainer): String? {
+    return try {
+      configurations.getByName(CLASSPATH_CONFIG)
+        .dependencies.first { it.name == SCABBARD_PLUGIN_ID }
+        .version.takeIf { it?.trim()?.isNotEmpty() == true }
+    } catch (e: Exception) {
+      null
+    }
+  }
+
+  fun calculate(): String {
+    val rootProjectConfigs = project.rootProject.buildscript.configurations
+    val currentProjectConfigs = project.buildscript.configurations
+    val versionFromRootProject = versionFromConfigurations(rootProjectConfigs)
+    //.also { project.logger.quiet("versionFromRootProject:$it") }
+    val versionFromCurrentProject = versionFromConfigurations(currentProjectConfigs)
+    //.also { project.logger.quiet("versionFromCurrentProject:$it") }
+    return versionFromCurrentProject ?: versionFromRootProject ?: "0.1.0" // <- worst case
+  }
+}
 
 internal fun Project.manageScabbardProcessor(enabled: Boolean) {
   val scabbardVersion = VersionCalculator(project).calculate()
